refactor: improve error handling with update check
