+++ package.json
@@ -26,11 +26,13 @@
 		"@actions/exec": "^1.0.4",
 		"@actions/io": "^1.0.2",
 		"@actions/tool-cache": "^1.6.1",
-		"libnpm": "^3.0.1"
+		"libnpm": "^3.0.1",
+		"semver": "^7.3.4"
 	},
 	"devDependencies": {
 		"@types/jest": "^26.0.16",
 		"@types/node": "^14.14.10",
+		"@types/semver": "^7.3.4",
 		"@typescript-eslint/eslint-plugin": "^4.9.0",
 		"@typescript-eslint/parser": "^4.9.0",
 		"@zeit/ncc": "^0.22.3",
