+++ tests/cache.test.ts
@@ -2,6 +2,7 @@ import os from 'os';
 import { join } from 'path';
 import * as remoteCache from '@actions/cache';
 import * as toolCache from '@actions/tool-cache';
+
 import * as cache from '../src/cache';
 import * as utils from './utils';
